backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiohttp-retry==2.9.1
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.3.0
bcrypt==4.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
colorama==0.4.6
Django==5.2.1
django-cors-headers==4.7.0
django-formtools==2.5.1
django-otp==1.6.0
django-phonenumber-field==8.1.0
django-two-factor-auth==1.17.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
frozenlist==1.7.0
idna==3.10
multidict==6.5.0
mysqlclient==2.2.7
phonenumbers==9.0.5
pillow==11.2.1
propcache==0.3.2
psycopg2-binary==2.9.10
PyJWT==2.9.0
pyotp==2.9.0
pypng==0.20220715.0
qrcode==7.4.2
requests==2.32.4
sqlparse==0.5.3
twilio==9.6.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.5.0
yarl==1.20.1
